(GH-545) Raise if identity is used with older git
The SSH implementation in this module relies on the GIT_SSH_COMMAND
environment variable. This option is only available in versions 2.10.0
and later of Git.

Therefore in in certain cases when consumers have an older version of
git installed the module will fail but would lack a descriptive error
message.

This commit ensure that we check the git version when the identity
parameter is used and raising a descriptive error if the current version
is lower than 2.10.0.