(GH-585/CONT-998) Fix for safe_directory logic
Due to  mistake in the logic unsafe directory was previously removed on every other run.
The logic previously checked whether the unsafe directory needed to be added, removing it if this was false without taking into account when it was already set but we wanted it to be left in place.