Fix home dir detection on windows
Add windows to the list of supported Operating Systems so that
acceptance tests are also run on this platform.

We need to set a clear-text password for the windows user provider
to avoid an OLE error code:800708C5 in Active Directory (The password
does not meet the password policy requirements).